feat: Add Validate Form (#9)
# This PR adds the Portfolio Project Validate Form following the design
and layout in the [Assignment
Link](https://github.com/bolah2009/curriculum-html-css/blob/main/portfolio/9_validate_form.md)
and the [Figma
Design](https://www.figma.com/file/LKueaSlr8Kc2yMKeZkKgtm/portfolio_project?type=design&node-id=48-2228&mode=design&t=SrlYt6mk7ReAfYmd-0)

This milestone is used to implement a client-side validation of the data
in the contact form before submission. and when the user submits the
form, it checks if the email is in lowercase.